backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
numpy>=1.24.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.1
langgraph>=0.0.10
python-dotenv>=1.0.0
requests>=2.31.0
mangum>=0.17.0